app/requirements.txt
geopandas==1.1.1
matplotlib==3.10.6
openeo==0.45.0
pyyaml==6.0.2
rasterio==1.4.3
streamlit-folium==0.25.2
tifffile==2025.6.11
timm==1.0.19
torch==2.7.1
